api/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pyspark
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2
python-multipart
pandas
pyarrow
numpy
python-dotenv